% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{agent}
\alias{agent}
\title{Define an agent}
\usage{
agent(id = NULL, type = "agent", location_y = NULL, location_x = NULL)
}
\arguments{
\item{id}{An ID for the agent. Can be a number or character name.}

\item{type}{Character string describing the agent's type e.g. 'predator' or 'prey'}

\item{location_y}{Agent's starting y coordinate}

\item{location_x}{Agent's starting x coordinate}
}
\value{
An 'agent' S3 object
}
\description{
Define an agent
}
\examples{
new_agent <- agent(id = 1)
}
