% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{agent_sim}
\alias{agent_sim}
\title{Set up an agentSim simulation}
\usage{
agent_sim(agents = NULL, time_total = NULL)
}
\arguments{
\item{time_total}{Number of time ticks to simulate}

\item{agent_population}{A population of agents}
}
\value{
a simulation
}
\description{
Set up an agentSim simulation
}
\examples{
#Build an example simulation in which each agent moves and then prints its coordinates to the console in each time step

#A parameter table to define 3 agents placed at random coordinates
init_pop <- tibble::tibble(
  id = c(1:3),
  type = "agent",
  location_x = runif(3, 0, 1),
  location_y = runif(3, 0, 1)
  )

#Define the task function, which will be run in each time step.
#You must pass 'agent' and 'population' as arguments to the function
#so that the agent can see itself and other agents when it runs its task.
move_and_report <- function(agent, population){

  #Move the agent
  agent$location_y <- agent$location_y + runif(1, -0.5, 0.5)
  agent$location_x <- agent$location_x + runif(1, -0.5, 0.5)

  #Print the agent's position to the console
  print(
    paste0("Agent '", agent$id, "' position: ",
    agent$location_y, ", ", agent$location_x)
  )

  agent
}

#Set up the simulation
simulation <- makePopulation(init_pop) \%>\%
  setTask(move_and_report) \%>\%
  agentSim(10)
}
