% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{agents_from_param_table}
\alias{agents_from_param_table}
\title{Creates a set of agents from a parameter table}
\usage{
agents_from_param_table(param_table)
}
\arguments{
\item{param_table}{Data frame with one row per agent. Each column will become
a named property of the created agents.}
}
\value{
list of agents
}
\description{
Creates a set of agents from a parameter table
}
