% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{iterate_sim}
\alias{iterate_sim}
\title{Run the next time step in a simulation}
\usage{
iterate_sim(x)
}
\arguments{
\item{x}{A simulation}
}
\value{
A simulation
}
\description{
Usually only called as part of runSim()
}
