% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Make a population of individual agents}
\usage{
population(agents)
}
\arguments{
\item{init_pop}{A table of parameters describing agents (one row per agent)}
}
\value{
A population
}
\description{
Takes a table of parameters describing a population and creates
the individual agents required.
}
\examples{
#A parameter table to define 3 agents placed at random coordinates
init_pop <- tibble::tibble(
  id = c(1:3),
  type = "agent",
  location_x = runif(3, 0, 1),
  location_y = runif(3, 0, 1)
  )

population <- makePopulation(init_pop)
}
