% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee.R
\name{referee}
\alias{referee}
\title{Define the referee}
\usage{
referee()
}
\value{
A 'referee' S3 object
}
\description{
The referee is a special class of agent with no parameters, whose function runs
after agents have completed their own functions in each time step.
You can think of the referee as being the environment's agent.
E.g. when simulating a game of soccer, the agents will be players and the environment
holds values such as the score. The referee is needed to increment the score and to reset
play after a team scores a goal.

Agents themselves could be given the referee's tasks but it's simpler to have a separate function that
runs at the end of each time step, after all of the agents have completed their tasks.

The referee has access to the entire simulation object and can modify any part of it,
including agents, the environment and the referee itself.
}
\examples{
new_referee <- referee()
}
