% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_sim}
\alias{run_sim}
\title{Run a simulation}
\usage{
run_sim(x, debug = FALSE)
}
\arguments{
\item{x}{A simulation created with agent_sim()}

\item{debug}{Optionally, print calculation step info to the console to aid in
debugging task errors}
}
\value{
The original simulation, with a populated simulation$history$log
}
\description{
Run a simulation
}
\examples{
#Example simulation in which each agent prints its own id to the console
init_pop <- tibble::tibble(
  id = c(1:50),
  x = runif(50, 0, 1),
  y = runif(50, 0, 1)
  )

agent_task <- function(agent, population){
  print(agent$id)
  agent
  }

sim <- agents_from_param_table(init_pop) \%>\%
  set_task(agent_task) \%>\%
  agent_sim(100) \%>\%
  run_sim()
}
