% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{set_task}
\alias{set_task}
\title{Set the task for an agent or a list of agents}
\usage{
set_task(x, task, agent_type = "agent")
}
\arguments{
\item{x}{An agent or a list of agents}

\item{task}{A function to be executed in each time step.
You must pass \code{agent} and \code{population} as arguments to the function
(see example below)}
}
\value{
The original agent or agent list 'x', with tasks appended
}
\description{
Set the task for a single agent, or for a list of agents,
which will be executed for each time step when the simulation is run.
The task function must take the arguments \code{agent} and \code{population}
and must return \code{agent}
}
\examples{
#Build an example simulation in which each agent moves and then prints its coordinates to the console in each time step
}
