% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{setLocation}
\alias{setLocation}
\title{Set the agent's location using x,y coordinates}
\usage{
setLocation(x, newLocation)
}
\arguments{
\item{x}{Agent}

\item{newLocation}{Location vector as a y,x (lon, lat) pair}
}
\value{
Agent
}
\description{
Set the agent's location using x,y coordinates
}
\examples{
new_agent <- agent(id = 1, location_y = 1, location_x = 2)

#Move the agent to coordinates y = 2, x = 3
new_agent <- setLocation(new_agent, c(2,3))
}
