% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{sim_environment}
\alias{sim_environment}
\title{Define a sim environment}
\usage{
sim_environment(params)
}
\arguments{
\item{params}{Named list of parameters that define the environment}
}
\value{
An 'environment' S3 object
}
\description{
Define a sim environment
}
\examples{
new_environment <- sim_environment(weather = 'rain')
}
