% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{agent}
\alias{agent}
\title{Define an agent}
\usage{
agent(params)
}
\arguments{
\item{params}{Named list of parameters that define the agent}
}
\value{
An 'agent' S3 object
}
\description{
Define an agent
}
\examples{
new_agent <- agent(id = 1)
}
