% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{agent_sim}
\alias{agent_sim}
\title{Set up an agentSim simulation}
\usage{
agent_sim(
  agents = NULL,
  time_total = NULL,
  environment = list(),
  referee = NULL
)
}
\arguments{
\item{agents}{}

\item{time_total}{Number of time ticks to simulate}
}
\value{
a simulation
}
\description{
Set up an agentSim simulation
}
\examples{
#Example simulation in which each agent prints its own id to the console
init_pop <- tibble::tibble(
  id = c(1:50),
  x = runif(50, 0, 1),
  y = runif(50, 0, 1)
  )

agent_task <- create_agent_task({
  print(agent$id)
  })

sim <- agents_from_param_table(init_pop) \%>\%
  set_task(agent_task) \%>\%
  agent_sim(10)
}
