% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{agents_from_param_table}
\alias{agents_from_param_table}
\title{Creates a set of agents from a parameter table}
\usage{
agents_from_param_table(param_table)
}
\arguments{
\item{param_table}{Data frame with one row per agent. Each column will become
a named property of the created agents.}
}
\value{
list of agents
}
\description{
Creates a set of agents from a parameter table
}
\examples{
init_pop <- tibble::tibble(
  id = c(1:50),
  x = runif(50, 0, 1),
  y = runif(50, 0, 1)
  )

new_agents <- agents_from_param_table(init_pop)
}
