% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{create_agent_task}
\alias{create_agent_task}
\title{Create a task to be executed by an agent(s)}
\usage{
create_agent_task(f)
}
\value{
A function with (agent, population) as pre-defined inputs and that returns agent
}
\description{
Create a task to be executed by an agent(s)
}
\examples{
agent_task <- create_agent_task({
  print(agent$id)
})
}
